er_triage/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0